📝 Add ColorPalette documentation and update README
