+++ .github/workflows/fastlane.yml
@@ -14,9 +14,14 @@ jobs:
       uses: actions/checkout@v1
     - name: Install dependencies
       run: |
+        pod deintegrate
         pod install --repo-update
       shell: bash
     - name: Bundle Install
       run: bundle install
-    - name: Unit tests
-      run: bundle exec fastlane sanity_checks
+    - name: Get the simulator up
+      run: |
+        fastlane run reset_simulator_contents
+        xcrun instruments -w 'iPhone SE (2nd generation)' || sleep 15
+    - name: Sanity tests
+      run: bundle exec fastlane sanity_checks --verbose
