+++ Gemfile
@@ -1,5 +1,4 @@
 source "https://rubygems.org"
 
 gem "cocoapods"
-gem "xcov"
 gem "fastlane"
