+++ SwiftyPick/Presentation/Screen/es.lproj/LaunchScreen.strings
