--- SwiftyPick/SceneDelegate.swift
@@ -1,27 +1,0 @@
-//
-//  SceneDelegate.swift
-//  SwiftyPick
-//
-//  Created by Manu Herrera on 24/08/2021.
-//
-
-import UIKit
-
-class SceneDelegate: UIResponder, UIWindowSceneDelegate {
-
-    var window: UIWindow?
-
-    func scene(
-        _ scene: UIScene,
-        willConnectTo session: UISceneSession,
-        options connectionOptions: UIScene.ConnectionOptions
-    ) {
-        guard let windowScene = (scene as? UIWindowScene) else { return }
-
-        let window = UIWindow(windowScene: windowScene)
-        window.rootViewController = ViewController()
-        self.window = window
-        window.makeKeyAndVisible()
-    }
-
-}
