+++ SwiftyPickUITests/SwiftyPickUITests.swift
@@ -8,19 +8,13 @@
 import XCTest
 
 class SwiftyPickUITests: XCTestCase {
-
-    func testExample() throws {
+    func testExample() {
         // UI tests must launch the application that they test.
         let app = XCUIApplication()
         app.launch()
 
+        return
         // Use recording to get started writing UI tests.
         // Use XCTAssert and related functions to verify your tests produce the correct results.
-    }
-
-    func testLaunchPerformance() throws {
-        measure(metrics: [XCTApplicationLaunchMetric()]) {
-            XCUIApplication().launch()
-        }
     }
 }
