--- fastlane/xcov_report/report.md
@@ -1,12 +1,0 @@
-## Current coverage for SwiftyPick.app is `37.50%`
-Files changed | - | - 
---- | --- | ---
-Assets+Generated.swift | `5.66%` | :skull:
-AppDelegate.swift | `44.44%` | :no_entry_sign:
-SceneDelegate.swift | `100.00%` | :white_check_mark:
-Strings+Generated.swift | `100.00%` | :white_check_mark:
-ViewController.swift | `100.00%` | :white_check_mark:
-
----
-
-> Powered by [xcov](https://github.com/nakiostudio/xcov)
